# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "tdPSA API Documentation",
        "description": "API documentation for tdPSA Clients module",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "tdPSA API Server"
        }
    ],
    "paths": {
...
        "/api/v1/email/inbox/messages": {
            "get": {
                "tags": [
                    "Email"
                ],
                "summary": "Get inbox messages",
                "description": "Returns paginated unrouted inbox messages. Messages already linked to tickets are excluded.",
                "operationId": "getInboxMessageList",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "account_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from_email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Missing email.read scope"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/email/inbox/messages/{message}": {
            "get": {
                "tags": [
                    "Email"
                ],
                "summary": "Get inbox message",
                "operationId": "getInboxMessageById",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Missing email.read scope"
                    },
                    "404": {
                        "description": "Inbox message not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/email/inbox/messages/{message}/spam": {
            "post": {
                "tags": [
                    "Email"
                ],
                "summary": "Mark inbox message as spam",
                "operationId": "markInboxMessageSpam",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message marked as spam"
                    },
                    "403": {
                        "description": "Missing email.update scope"
                    },
                    "404": {
                        "description": "Inbox message not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/email/inbox/poll": {
            "post": {
                "tags": [
                    "Email"
                ],
                "summary": "Queue inbox polling",
                "operationId": "queueInboxPoll",
                "responses": {
                    "202": {
                        "description": "Inbox polling queued"
                    },
                    "403": {
                        "description": "Missing email.update scope"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}